+++ packages/core/src/ir/index.ts
@@ -1,0 +1,4 @@
+export * from './shared.js'
+export * from './requirement.js'
+export * from './design.js'
+export * from './implementation.js'